env-data-pipeline/requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
numpy==2.0.2
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
requests==2.32.5
six==1.17.0
tzdata==2025.3
urllib3==1.26.20
